Move validation functions into separate files
