+++ index.js
@@ -12,7 +12,8 @@ Object.assign(global, {
 
 import setupBoard from './src/board/setup-board.js';
 import createBoard from './src/board/create-board.js';
-import * as validation from './src/validation.js';
+import * as validation from './src/validation/validation.js';
+import makeMove from './src/validation/make-move.js';
 import findAllMoves from './src/validation/all-moves.js';
 import isCheck from './src/validation/is-check.js';
 import gameEndingStatus from './src/validation/ending.js';
@@ -21,6 +22,7 @@ import undoMove from './src/board/undo.js';
 export default {
 	setupBoard,
 	createBoard,
+	makeMove,
 	validation,
 	findAllMoves,
 	isCheck,
