+++ src/eval/material-diff.js
@@ -5,7 +5,7 @@ export default function materialDifference() {
 		for (let j = 0; j < i.length; i++) {
 			const colour = i[j] === i[j].toUpperCase() ? 'w' : 'b';
 			const piece = i[j].toLowerCase();
-			let pointsIncrease = { p: 1, m: 3, b: 3, r: 5, q: 9 }[piece];
+			const pointsIncrease = { p: 1, m: 3, b: 3, r: 5, q: 9 }[piece];
 			points[colour] += pointsIncrease;
 		}
 	}
